boards/arm/stm32h7: switch board common to boards/arm/common/stm32
Wire the STM32H7 family to the shared boards/arm/common/stm32 board-common
directory.

Drop the local copies that are already provided by the shared directory:

  - stm32_reset.c from nucleo-h723zg, nucleo-h743zi, nucleo-h743zi2,
    nucleo-h753zi, stm32h745i-disco and stm32h750b-dk
  - stm32_romfs_initialize.c and stm32_romfs.h from nucleo-h743zi (identical
    to the shared versions)

Enable CONFIG_ARCH_BOARD_COMMON=y in the configs that use CONFIG_BOARDCTL_RESET
or CONFIG_STM32_ROMFS so board_reset()/romfs are pulled from the shared
directory.
