+++ boards/arm/stm32h7/linum-stm32h753bi/configs/qencoder/defconfig
@@ -10,6 +10,7 @@
 # CONFIG_STANDARD_SERIAL is not set
 CONFIG_ARCH="arm"
 CONFIG_ARCH_BOARD="linum-stm32h753bi"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD_LINUM_STM32H753BI=y
 CONFIG_ARCH_CHIP="stm32h7"
 CONFIG_ARCH_CHIP_STM32=y
