+++ boards/arm/stm32h7/linum-stm32h753bi/src/linum-stm32h753bi.h
@@ -106,11 +106,6 @@
 #define BUZZER_PWMTIMER    4
 #define BUZZER_PWMCHANNEL  2
 
-/* OneShot Timer */
-
-#define BOARD_TONE_ONESHOT_TIM     17  /* Timer 17 - Oneshot timer for note timings */
-#define BOARD_TONE_ONESHOT_TIM_RES 10  /* Timer 17 - Oneshot timer resolution (us)  */
-
 /* Ethernet
  *
  * PI4  Reset PHY pin
@@ -242,16 +237,6 @@ int stm32_w25qxxx_setup(void);
 #endif
 
 /****************************************************************************
- * Name: board_qencoder_initialize
- *
- * Description:
- *   Initialize the quadrature encoder driver for the given timer
- *
- ****************************************************************************/
-
-int board_qencoder_initialize(int devno, int timerno);
-
-/****************************************************************************
  * Name: stm32_mfrc522initialize
  *
  * Description:
@@ -267,21 +252,6 @@ int board_qencoder_initialize(int devno, int timerno);
 
 #ifdef CONFIG_CL_MFRC522
 int stm32_mfrc522initialize(const char *devpath);
-#endif
-
-/****************************************************************************
- * Name: board_tone_initialize
- *
- * Input Parameters:
- *   devno - The device number, used to build the device path as /dev/toneN
- *
- * Description:
- *   Configure and initialize the tone generator.
- *
- ****************************************************************************/
-
-#ifdef CONFIG_AUDIO_TONE
-int board_tone_initialize(int devno);
 #endif
 
 /****************************************************************************
