+++ boards/arm/stm32h7/linum-stm32h753bi/src/stm32_bringup.c
@@ -38,6 +38,14 @@
 
 #include "linum-stm32h753bi.h"
 
+#ifdef CONFIG_SENSORS_QENCODER
+#  include "board_qencoder.h"
+#endif
+
+#ifdef CONFIG_AUDIO_TONE
+#  include "stm32_tone.h"
+#endif
+
 #ifdef CONFIG_USERLED
 #include <nuttx/leds/userled.h>
 #endif
