+++ boards/arm/stm32h7/nucleo-h723zg/src/Make.defs
@@ -1,5 +1,5 @@
 ############################################################################
-# boards/arm/stm32h7/weact-stm32h750/src/Makefile
+# boards/arm/stm32h7/nucleo-h723zg/src/Make.defs
 #
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -22,7 +22,11 @@
 
 include $(TOPDIR)/Make.defs
 
-CSRCS = stm32_boot.c stm32_bringup.c stm32_usb.c stm32_ioctl.c stm32_spi.c
+CSRCS = stm32_boot.c stm32_bringup.c
+
+ifeq ($(CONFIG_ADC),y)
+CSRCS += stm32_adc.c
+endif
 
 ifeq ($(CONFIG_ARCH_LEDS),y)
 CSRCS += stm32_autoleds.c
@@ -30,18 +34,34 @@ else
 CSRCS += stm32_userleds.c
 endif
 
-ifeq ($(CONFIG_VIDEO_FB),y)
-  ifeq ($(CONFIG_LCD_ST7735),y)
-    CSRCS += stm32_lcd_st7735.c
-  endif
+ifeq ($(CONFIG_ARCH_BUTTONS),y)
+CSRCS += stm32_buttons.c
 endif
 
-ifeq ($(CONFIG_STM32_SDMMC),y)
-CSRCS += stm32_sdmmc.c
+ifeq ($(CONFIG_STM32_SPI),y)
+CSRCS += stm32_spi.c
 endif
 
-ifeq ($(CONFIG_FAT_DMAMEMORY),y)
-CSRCS += stm32_dma_alloc.c
+ifeq ($(CONFIG_STM32_OTGHS),y)
+CSRCS += stm32_usb.c
 endif
 
-include $(TOPDIR)/boards/Board.mk
+ifeq ($(CONFIG_BOARDCTL_UNIQUEID),y)
+CSRCS += stm32_uid.c
+endif
+
+ifeq ($(CONFIG_DEV_GPIO),y)
+CSRCS += stm32_gpio.c
+endif
+
+ifeq ($(CONFIG_PWM),y)
+CSRCS += stm32_pwm.c
+endif
+
+ifeq ($(CONFIG_NET_OA_TC6),y)
+CSRCS += stm32_oa_tc6.c
+endif
+
+DEPPATH += --dep-path board
+VPATH += :board
+CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)board
