+++ boards/arm/stm32h7/nucleo-h743zi/configs/elf/defconfig
@@ -11,6 +11,7 @@
 # CONFIG_STM32_DTCMEXCLUDE is not set
 CONFIG_ARCH="arm"
 CONFIG_ARCH_BOARD="nucleo-h743zi"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD_NUCLEO_H743ZI=y
 CONFIG_ARCH_CHIP="stm32h7"
 CONFIG_ARCH_CHIP_STM32=y
