+++ boards/arm/stm32h7/nucleo-h743zi/configs/nxboot-app/defconfig
@@ -8,6 +8,7 @@
 # CONFIG_STANDARD_SERIAL is not set
 CONFIG_ARCH="arm"
 CONFIG_ARCH_BOARD="nucleo-h743zi"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD_NUCLEO_H743ZI=y
 CONFIG_ARCH_CHIP="stm32h7"
 CONFIG_ARCH_CHIP_STM32=y
