+++ boards/arm/stm32h7/nucleo-h743zi/configs/nxboot-loader/defconfig
@@ -7,6 +7,7 @@
 #
 CONFIG_ARCH="arm"
 CONFIG_ARCH_BOARD="nucleo-h743zi"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD_NUCLEO_H743ZI=y
 CONFIG_ARCH_CHIP="stm32h7"
 CONFIG_ARCH_CHIP_STM32=y
