+++ boards/arm/stm32h7/nucleo-h743zi2/configs/jumbo/defconfig
@@ -9,6 +9,7 @@
 # CONFIG_STANDARD_SERIAL is not set
 CONFIG_ARCH="arm"
 CONFIG_ARCH_BOARD="nucleo-h743zi2"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD_NUCLEO_H743ZI2=y
 CONFIG_ARCH_BUTTONS=y
 CONFIG_ARCH_CHIP="stm32h7"
