+++ boards/arm/stm32h7/nucleo-h753zi/src/Make.defs
@@ -1,5 +1,7 @@
 ############################################################################
-# boards/arm/stm32h7/stm32h750b-dk/src/Makefile
+# boards/arm/stm32h7/nucleo-h753zi/src/Make.defs
+#
+# SPDX-License-Identifier: Apache-2.0
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
@@ -22,30 +24,36 @@ include $(TOPDIR)/Make.defs
 
 CSRCS = stm32_boot.c stm32_bringup.c
 
+ifeq ($(CONFIG_ADC),y)
+CSRCS += stm32_adc.c
+endif
+
 ifeq ($(CONFIG_ARCH_LEDS),y)
 CSRCS += stm32_autoleds.c
 else
 CSRCS += stm32_userleds.c
 endif
 
+ifeq ($(CONFIG_ARCH_BUTTONS),y)
+CSRCS += stm32_buttons.c
+endif
+
 ifeq ($(CONFIG_STM32_OTGFS),y)
 CSRCS += stm32_usb.c
 endif
 
-ifeq ($(CONFIG_TESTING_OSTEST),y)
-CSRCS += stm32_ostest.c
+ifeq ($(CONFIG_DEV_GPIO),y)
+CSRCS += stm32_gpio.c
 endif
 
-ifeq ($(CONFIG_BOARDCTL_RESET),y)
-CSRCS += stm32_reset.c
+ifeq ($(CONFIG_PWM),y)
+CSRCS += stm32_pwm.c
 endif
 
-ifeq ($(CONFIG_INPUT_FT5X06),y)
-CSRCS += stm32_ft5x06.c
+ifeq ($(CONFIG_SENSORS_QENCODER),y)
+CSRCS += stm32_qencoder.c
 endif
 
-ifeq ($(CONFIG_STM32_LTDC),y)
-CSRCS += stm32_lcd.c
-endif
-
-include $(TOPDIR)/boards/Board.mk
+DEPPATH += --dep-path board
+VPATH += :board
+CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)board
