+++ boards/arm/stm32h7/stm32h750b-dk/src/CMakeLists.txt
@@ -34,10 +34,6 @@ if(CONFIG_TESTING_OSTEST)
   list(APPEND SRCS stm32_ostest.c)
 endif()
 
-if(CONFIG_BOARDCTL_RESET)
-  list(APPEND SRCS stm32_reset.c)
-endif()
-
 if(CONFIG_INPUT_FT5X06)
   list(APPEND SRCS stm32_ft5x06.c)
 endif()
